EPlusX/requirements.txt
pdfplumber
langchain
openai
hnswlib
numpy
pandas
openpyxl
python-dotenv